+++ next-docs/pages/components/loader.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
 import Preview from '../../components/codePreview/Preview';
-import Customization from '../../public/examples/loader/Customization';
+import Table from '../../components/Table';
+import Colours from '../../public/examples/loader/Colours';
 import Default from '../../public/examples/loader/Default';
+import Sizes from '../../public/examples/loader/Sizes';
 import useExamples from '../../utils/useExamples';
 
 export default function PageLoader() {
@@ -11,10 +13,17 @@ export default function PageLoader() {
     <>
       <section className="mt-8">
         <h1 className="text-5xl font-semibold">Loader</h1>
-        <p className="text-lg mt-4">Pure CSS loading animation.</p>
+        <p className="text-lg mt-4">
+          Fondly nicknamed “the launcher”, the Loader ensures users that
+          progress is happening so they don't give up and leave the rocket page.
+        </p>
+        <p className="text-lg mt-4">
+          The Loader, along with a succinct message, is especially important
+          after a user initiates an action that requires them to wait. This is
+          so that they know there's no need to repeat what they've done.
+        </p>
       </section>
 
-      {/* Default */}
       <section className="mt-8">
         <Preview
           title="Default"
@@ -23,12 +32,41 @@ export default function PageLoader() {
         />
       </section>
 
-      {/* Customization */}
       <section className="mt-8">
         <Preview
           title="Colours"
-          preview={<Customization />}
-          code={examples ? examples.Customization : 'Loading'}
+          preview={<Colours />}
+          code={examples ? examples.Colours : 'Loading'}
+        />
+      </section>
+
+      <section className="mt-8">
+        <Preview
+          title="Sizes"
+          preview={<Sizes />}
+          code={examples ? examples.Sizes : 'Loading'}
+        />
+      </section>
+
+      <section className="mt-14">
+        <h2 className="text-2xl mb-4">Props</h2>
+        <Table
+          data={[
+            {
+              name: 'color',
+              type: 'ColorProps',
+              required: false,
+              default: 'bulma',
+              description: 'Color of the component',
+            },
+            {
+              name: 'size',
+              type: 'twoxsmall | xsmall | small | medium | large',
+              required: false,
+              default: 'medium',
+              description: 'Size of the component',
+            },
+          ]}
         />
       </section>
     </>