+++ next-docs/public/examples/chip/Sizes.tsx
@@ -1,10 +1,9 @@
-import { Chip } from '@heathmont/moon-components';
+import { Chip } from '@heathmont/moon-core';
 
 const Example = () => (
   <div className="flex justify-around items-center w-full">
     <Chip size="small">Small</Chip>
-    <Chip size="medium">Medium</Chip>
-    <Chip>Default</Chip>
+    <Chip>Medium is Default</Chip>
   </div>
 );