+++ next-docs/public/examples/heading/Colors.tsx
@@ -1,10 +1,11 @@
-import React from 'react';
 import { Heading } from '@heathmont/moon-core';
 
 const Example = () => (
-  <div className="flex justify-around items-center w-full">
-    <Heading color={'goku.80'}>Title</Heading>
-    <Heading color={'hit.80'}>Title</Heading>
+  <div className="flex justify-around w-full items-center">
+    <Heading>Defaul color is Bulma</Heading>
+    <Heading color="piccolo.100">Piccolo</Heading>
+    <Heading color="trunks.100">Trunks</Heading>
+    <Heading color="krillin.100">Krillin</Heading>
   </div>
 );