+++ next-docs/public/examples/loader/__tests__/index.test.tsx
@@ -6,7 +6,8 @@ import React from 'react';
 import { moonDesignLight, ThemeProvider } from '@heathmont/moon-themes';
 import renderer from 'react-test-renderer';
 import Default from '../Default';
-import Customization from '../Customization';
+import Colours from '../Colours';
+import Sizes from '../Sizes';
 
 const renderWithTheme = (component: JSX.Element) => (
   <ThemeProvider theme={moonDesignLight}>{component}</ThemeProvider>
@@ -19,7 +20,12 @@ describe('Loader', () => {
   });
 
   it('renders with different colours', () => {
-    const testRenderer = renderer.create(renderWithTheme(<Customization />));
+    const testRenderer = renderer.create(renderWithTheme(<Colours />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders with sizes', () => {
+    const testRenderer = renderer.create(renderWithTheme(<Sizes />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 });