+++ next-docs/public/examples/text/Size.tsx
@@ -1,5 +1,0 @@
-import { Text } from '@heathmont/moon-core';
-
-const Example = () => <Text size={64} />;
-
-export default Example;