+++ packages/core/src/banner/styles/BodyButtonContainer.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const BodyButtonContainer = styled.div(({ theme: { newTokens } }) => ({
+  display: 'flex',
+  justifyContent: 'end',
+  gap: newTokens.space.twoxsmall,
+  alignItems: 'center',
+}));
+
+export default BodyButtonContainer;