+++ packages/core/src/banner/styles/Content.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const Content = styled.div({
+  width: '100%',
+});
+
+export default Content;