+++ packages/core/src/banner/styles/HeaderIcon.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const HeaderIcon = styled.div({
+  display: 'flex',
+  justifyContent: 'end',
+  cursor: 'pointer',
+});
+
+export default HeaderIcon;