+++ packages/select/src/private/types/TextInputGroupType.tsx
@@ -1,0 +1,1 @@
+export type TextInputGroupTypes = 'single-select' | 'multi-select';