+++ next-docs/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Change Log
 
+## 8.3.0
+
+### Minor Changes
+
+- Fix TextInput Groups
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@8.3.0
+  - @heathmont/moon-components@8.3.0
+  - @heathmont/moon-core@8.3.0
+  - @heathmont/moon-select@8.3.0
+  - @heathmont/moon-accordion@8.3.0
+  - @heathmont/moon-charts@8.3.0
+  - @heathmont/moon-datepicker@8.3.0
+  - @heathmont/moon-draggabletable@8.3.0
+  - @heathmont/moon-icons@8.3.0
+  - @heathmont/moon-pagination@8.3.0
+  - @heathmont/moon-popover@8.3.0
+  - @heathmont/moon-sidebar@8.3.0
+  - @heathmont/moon-table@8.3.0
+  - @heathmont/moon-themes@8.3.0
+  - @heathmont/moon-utils@8.3.0
+
 ## 8.2.0
 
 ### Minor Changes