+++ packages/fonts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 8.3.0
+
+### Minor Changes
+
+- Fix TextInput Groups
+
 ## 8.2.0
 
 ### Minor Changes